SSH: when we add a new processing server, in the afteradd method, we'll retrieve a path( this path with locate the place where we want to stock the ssh keys) and then, we'll call the method for creating a pair of ssh key.
Method  createKPairSSH(String keyPath,String hostname): this method will create a folder in the location of the keypath with the name of the hostname. In it, we'll create a pair of ssh key.
Of course, this method check severals things: if the directory already exists, if the files already exist,...